Fixed format specifiers and some casts to fix trace output

+++ iodbc/trace/trace.c
@@ -717,7 +717,7 @@ _trace_print_function (int func, int trace_leave, int retcode)
       tv.tv_sec--;
       tv.tv_usec += 1000000L;
     }
-  trace_emit ("\n[%06ld.%06ld]\n", tv.tv_sec, tv.tv_usec);
+  trace_emit ("\n[%06ld.%06ld]\n", (long) tv.tv_sec, (long) tv.tv_usec);
 #else
   trace_emit ("\n");
 #endif
@@ -1316,7 +1316,7 @@ _trace_bufferlen (SQLINTEGER length)
     }
 
   if (ptr)
-    trace_emit ("\t\t%-15.15s * %ld (%s)\n", "SQLINTEGER", length, ptr);
+    trace_emit ("\t\t%-15.15s * %ld (%s)\n", "SQLINTEGER", (long) length, ptr);
   else
     trace_emit ("\t\t%-15.15s * %ld\n", "SQLINTEGER", (long) length);
 }
